+++ home-work-lesson-9/src/main/java/ru/otus/spring/course/entities/Author.java
@@ -5,13 +5,10 @@ import lombok.experimental.Accessors;
 
 import javax.persistence.*;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-@Getter
-@Setter
-@NoArgsConstructor
+@Data
 @Entity(name = "author")
 @Table(name = "author")
 @Accessors(chain = true)
@@ -24,31 +21,6 @@ public class Author {
     @Column(name = "surname")
     private String surname;
 
-    @ManyToMany(mappedBy = "authors", fetch = FetchType.EAGER)
+    @ManyToMany(mappedBy = "authors")
     private Set<Book> books = new HashSet<>();
-
-    @Override
-    public String toString() {
-        return "Author{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", surname='" + surname + '\'' +
-                ", books=" + books +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Author)) return false;
-        Author author = (Author) o;
-        return getId().equals(author.getId()) &&
-                getName().equals(author.getName()) &&
-                getSurname().equals(author.getSurname());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getId(), getName(), getSurname());
-    }
 }
