+++ home-work-lesson-9/src/main/java/ru/otus/spring/course/entities/Comment.java
@@ -18,7 +18,7 @@ public class Comment {
     @Column(name = "text", nullable = false)
     private String text;
 
-    @ManyToOne(fetch = FetchType.EAGER)
+    @ManyToOne
     @JoinColumn(name = "id_book")
     private Book book;
 }
