+++ home-work-lesson-9/src/main/java/ru/otus/spring/course/repository/StyleRepositoryImpl.java
@@ -1,0 +1,43 @@
+package ru.otus.spring.course.repository;
+
+import org.springframework.stereotype.Repository;
+import ru.otus.spring.course.entities.Style;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import java.util.List;
+
+@Repository
+public class StyleRepositoryImpl implements StyleRepository {
+    @PersistenceContext
+    private EntityManager entityManager;
+
+    @Override
+    public Style findById(String id) {
+        return entityManager.find(Style.class, id);
+    }
+
+    @Override
+    public List<Style> findAll() {
+        return entityManager.createQuery("from style", Style.class).getResultList();
+    }
+
+    @Override
+    public void delete(Style style) {
+        entityManager.remove(style);
+    }
+
+    @Override
+    public Style save(Style style) {
+        if (style.getId() == null) {
+            entityManager.persist(style);
+            return style;
+        } else
+            return entityManager.merge(style);
+    }
+
+    @Override
+    public Style update(Style style) {
+        return entityManager.merge(style);
+    }
+}
