+++ home-work-lesson-9/src/test/java/ru/otus/spring/course/repository/StyleRepositoryTest.java
@@ -15,10 +15,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 @DataJpaTest
-@Import(StyleRepository.class)
+@Import(StyleRepositoryImpl.class)
 public class StyleRepositoryTest {
     @Autowired
-    private StyleRepository styleRepository;
+    private StyleRepositoryImpl styleRepository;
 
     @Autowired
     private TestEntityManager em;
